% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NewtonTriseccion.R
\name{triseccion}
\alias{triseccion}
\title{Busca una raiz con el metodo de triseccion}
\usage{
triseccion(a, b, f, tol, prec)
}
\arguments{
\item{a}{limite menor del intervalo}

\item{b}{limite mayor del intervalo}

\item{f}{ecuacion no lineal}

\item{tol}{tolerancia}

\item{prec}{precision de la respuesta}
}
\value{
la aproximacion a la raiz de f, si existe
}
\description{
Por medio de una particion en 3 intervalos de un intervalo a,b, iterativamente el programa busca una raiz para una ecuacion
no lineal
}
\examples{
r=raizENL(f,a,b,tol,prec);
r=raizENL(f,a,b,,prec);
r=raizENL(f,a,b);


}
