% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NewtonTriseccion.R
\name{newtonMejorado}
\alias{newtonMejorado}
\title{Busca una raiz con el metodo de newton mejorado}
\usage{
newtonMejorado(f, xi, tol, prec)
}
\arguments{
\item{f}{ecuacion no lineal}

\item{xi}{x inicial}

\item{tol}{tolerancia}

\item{prec}{precision de la respuesta}
}
\value{
la aproximacion a la raiz de f, si existe
}
\description{
Por medio del metodo de newton, busca una raiz aproximada de una ecuacion no lineal
}
\examples{
r=raizENL(f,xi,tol,prec);
r=raizENL(f,xi,tol);
r=raizENL(f,xi);


}
