% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Triseccion.R
\name{triseccion}
\alias{triseccion}
\title{Busca una raíz con el método de trisección}
\usage{
triseccion(f, a, b, tol = 1e-08, prec = 6)
}
\arguments{
\item{f}{función de una ecuación no lineal}

\item{a}{limite menor del intervalo}

\item{b}{limite mayor del intervalo}

\item{tol}{tolerancia (por defecto : 1e-8)}

\item{prec}{precisión de la respuesta (por defecto: 6)}
}
\value{
la aproximación a la raíz de f, si existe y el error
}
\description{
Por medio de una partición en 3 intervalos de un intervalo a,b, iterativamente el programa busca una raíz para una ecuación
no lineal
}
\examples{

f = function (x) x^2-3
a = 1
b = 2
tol = 1e-6
prec = 5
res = triseccion (f, a, b, tol, prec)
res

res = triseccion (f, a, b, tol)
res

res = triseccion (f, a, b)
res


}
